Refactoring out age transformation due to size.
